Tooltip improvements
Tooltips are now nicely uniformly rendered. All panels can now expose an
unlimited amount of tooltips. The UI will draw as many as fit on the
bottom tooltip bar, and if not all fit, then a ? key is enabled that
will show all functions in a text box in place of the current focused
panel.